src/backend/requirements.txt
fastapi==0.104.0
filetype==1.2.0
jq==1.6.0
langchain==0.0.316
numpy==1.24.3
openai==1.6.1
pandas==2.1.1
pgvector==0.2.3
psycopg2-binary==2.9.9
python-docx==1.1.0
python-multipart==0.0.6
tiktoken==0.5.1
uvicorn==0.23.2
haystack-ai==2.0.0b4
pypdf==3.17.4
huggingface_hub==0.20.2